requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-ckeditor==6.4.2
django-image-optimizer==1.0.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-translated-fields==0.12.0
idna==3.3
jdatetime==4.1.0
Pillow==9.2.0
python-resize-image==1.1.20
requests==2.28.1
sqlparse==0.4.2
tinify==1.5.2
urllib3==1.26.10
